moved GUI functions to GUI service
