+++ src/app/app.module.ts
@@ -1,6 +1,6 @@
 //angular imports
 import { BrowserModule } from '@angular/platform-browser';
-import { NgModule, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
+import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 //components
 import { AppRoutingModule } from './app-routing.module';
@@ -11,7 +11,6 @@ import { ClassBoxComponent } from './class-box/class-box.component';
 import { ClassAreaComponent } from './class-area/class-area.component';
 
 //services
-//import { StorageService } from './storage.service';
 import { ClassStorageService } from './class-storage.service'
 
 //material imports
