+++ src/app/class-area/class-area.component.html
@@ -2,14 +2,14 @@
     <mat-toolbar class="menu">
         <h3>GWCA</h3>
         <button mat-raised-button color="primary" class="buttons new" (click)="openDialog('new')" >New</button>
-        <button mat-raised-button color="primary" class="buttons edit" (click)="openDialog('edit');service.connectionsUpdateWrapper()">Edit</button>
+        <button mat-raised-button color="primary" class="buttons edit" (click)="openDialog('edit');guiService.connectionsUpdateWrapper()">Edit</button>
         <button mat-raised-button color="primary" [matMenuTriggerFor]="delete">Delete</button>
         <mat-menu #delete>
             <button mat-menu-item color="primary" (click)="openDialog('delete')">Delete Class</button>
             <button mat-menu-item color="primary" (click)="openDialog('delete_attribute')">Delete Attribute</button>
         </mat-menu>
         <button mat-raised-button color="primary" class="buttons import" (click)="openDialog('import')">Import</button>
-        <button mat-raised-button color="primary" class="buttons export" (click)="openDialog('export');updateStoredDiagram();service.connectionsUpdateWrapper()">Export</button>
+        <button mat-raised-button color="primary" class="buttons export" (click)="openDialog('export');updateStoredDiagram();guiService.connectionsUpdateWrapper()">Export</button>
         <button mat-raised-button color="primary" class="buttons help" (click)="openDialog('help')">Help</button>
         <div class="cli-container">
             <a routerLink="cli">
