+++ src/app/class-box/class-box.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit, AfterViewInit, IterableDiffer, IterableDiffers, DoCh
 import { ClassStorageService} from '../class-storage.service';
 import { MatDialog, MatDialogRef } from '@angular/material';
 import { DialogTestComponent } from '../dialog-test/dialog-test.component';
+import { GuiStorageService } from '../gui-storage.service';
 
 @Component({
   selector: 'app-class-box',
@@ -27,7 +28,7 @@ export class ClassBoxComponent implements OnInit, AfterViewInit,DoCheck, AfterVi
 
 
   constructor(public service: ClassStorageService, public dialog: MatDialog,
-    private iterableDiffs: IterableDiffers) { 
+    private iterableDiffs: IterableDiffers, public guiService: GuiStorageService) { 
       this.iterableDiffer= this.iterableDiffs.find([]).create(null);
 
   }
@@ -43,10 +44,10 @@ export class ClassBoxComponent implements OnInit, AfterViewInit,DoCheck, AfterVi
       }
     }
     this.name = this.service.allClasses[this.index]['name'];
-    this.id = this.name + "_" + this.service.instance;
+    this.id = this.name + "_" + this.guiService.instance;
     this.variables = this.service.allClasses[this.index]['variables'];
     this.methods = this.service.allClasses[this.index]['methods'];
-    this.service.instance = this.service.instance + 1;
+    this.guiService.instance = this.guiService.instance + 1;
 
 
 
@@ -64,7 +65,7 @@ export class ClassBoxComponent implements OnInit, AfterViewInit,DoCheck, AfterVi
   }
 
   ngAfterViewInit(){
-    var jsPlumbInstance = this.service.jsPlumbInstance;
+    var jsPlumbInstance = this.guiService.jsPlumbInstance;
        
 
      //get all the dynamically created elements
@@ -74,25 +75,25 @@ export class ClassBoxComponent implements OnInit, AfterViewInit,DoCheck, AfterVi
      //keep track of the top shift value
      var topShift  = 20;
      for(var i = 0;i<boxes.length;i++){
-      if(this.service.leftShift > 900){
-        this.service.leftShift = 20;
+      if(this.guiService.leftShift > 900){
+        this.guiService.leftShift = 20;
         topShift = topShift + 300;
       }
-        (<HTMLElement>boxes[i]).style.left = this.service.leftShift + 'px';
+        (<HTMLElement>boxes[i]).style.left = this.guiService.leftShift + 'px';
         (<HTMLElement>boxes[i]).style.top = topShift + 'px';
-        this.service.leftShift = this.service.leftShift + 300;
+        this.guiService.leftShift = this.guiService.leftShift + 300;
      }
 
-    var jsPlumbInstance = this.service.jsPlumbInstance;
+    var jsPlumbInstance = this.guiService.jsPlumbInstance;
     //note this is kind of a hacky solution...
 
-    this.service.addEndpoints(this.id);
+    this.guiService.addEndpoints(this.id);
 
     var id = this.id;
 
  
     setTimeout(() =>
-    this.service.jsPlumbInstance.draggable(this.id,{
+    this.guiService.jsPlumbInstance.draggable(this.id,{
       drag:function(event){
         jsPlumbInstance.revalidate(id);
         jsPlumbInstance.repaintEverything();
@@ -147,126 +148,11 @@ export class ClassBoxComponent implements OnInit, AfterViewInit,DoCheck, AfterVi
     this.autoConnect();
 
 
-    //no self connections
-    jsPlumbInstance.bind("connection",function(){
-      var connections = jsPlumbInstance.getConnections(this.id);
-      //remove self connections
-        if(connections[(connections.length - 1)]['source']['id'] == connections[(connections.length - 1)]['target']['id']){
-          jsPlumbInstance.deleteConnection(connections[(connections.length - 1)]);
-        }
-        //otherwise open dialog so we can specify connection type
-        else{
-          if(id == connections[(connections.length - 1)]['source']['id']){
-            dialogRef = dialog.open(DialogTestComponent, {width: '250px'});
-            dialogRef.componentInstance.buttonPressed = "connection";
-            dialogRef.afterClosed().subscribe(() => {
-              connectionType = dialogRef.componentInstance.connectionType;
-              switch(connectionType){
-                //open diamond, solid line
-                //aggregation
-                case 'purple':
-                  connections[(connections.length - 1)].removeAllOverlays();
-                  connections[(connections.length - 1)].setPaintStyle({stroke: 'purple', lineWidth: '10px'});
-                  connections[(connections.length - 1)].addOverlay(["Label",{label:"Aggregation",location:0.5}]);
-                  connections[(connections.length - 1)].addOverlay(["Diamond",{
-                    cssClass: "unfilledDiamond",
-                    width: 15,
-                    length: 30,
-                    location: 1
-                  }]);
-                  break;
-                //association
-                //just a line
-                case 'green':
-                  connections[(connections.length - 1)].removeAllOverlays();
-                  connections[(connections.length - 1)].setPaintStyle({stroke: 'green', lineWidth: '10px'});
-                  connections[(connections.length - 1)].addOverlay(["Label",{label:"Association",location:0.5}]);
-                  break;
-                //closed diamond, solid line
-                //composition
-                case 'red':
-                  connections[(connections.length - 1)].removeAllOverlays();
-                  connections[(connections.length - 1)].setPaintStyle({stroke: 'red', lineWidth: '10px'});
-                  connections[(connections.length - 1)].addOverlay(["Label",{label:"Composition",location:0.5}]);
-                  connections[(connections.length - 1)].addOverlay(["Diamond", {width: 15,length: 30,location: 1}]);
-                  break;
-                //closed arrow, solid line
-                //generalization
-                case 'orange':
-                  connections[(connections.length - 1)].setPaintStyle({stroke: 'orange', lineWidth: '10px'});
-                  connections[(connections.length - 1)].addOverlay(["Label",{label:"Generalization",location:0.5}]);
-                  break;
-                //closed arrow, dashed line
-                //realization
-                case 'yellow':
-                  connections[(connections.length - 1)].setPaintStyle({"dashstyle":'3 3',stroke:'yellow',strokeWidth:5});
-                  connections[(connections.length - 1)].addOverlay(["Label",{label:"Realization",location:0.5}]);
-                  break;
-              }
-            });
-          }
-        }
-
-      });
+    //bind everything 
+    this.guiService.bindConnections(dialogRef,dialog,jsPlumbInstance,connectionType,DialogTestComponent,id);
 
 
-    //on connection click
-    jsPlumbInstance.bind("click",function(connection){
-      if(id == connection['source']['id']){
-        dialogRef = dialog.open(DialogTestComponent, {width: '250px'});
-        dialogRef.componentInstance.buttonPressed = "connection";
-        dialogRef.afterClosed().subscribe(() => {
-          connectionType = dialogRef.componentInstance.connectionType;
-          switch(connectionType){
-            //open diamond, solid line
-            //aggregation
-            case 'purple':
-              connection.removeAllOverlays();
-              connection.setPaintStyle({stroke: 'purple', lineWidth: '10px'});
-              connection.addOverlay(["Label",{label:"Aggregation",location:0.5}]);
-              connection.addOverlay(["Diamond",{
-                cssClass: "unfilledDiamond",
-                width: 15,
-                length: 30,
-                location: 1
-              }]);
-
-              break;
-            //association
-            //just a line
-            case 'green':
-              connection.removeAllOverlays();
-              connection.setPaintStyle({stroke: 'green', lineWidth: '10px'});
-              connection.addOverlay(["Label",{label:"Association",location:0.5}]);
-              break;
-            //closed diamond, solid line
-            //composition
-            case 'red':
-              connection.removeAllOverlays();
-              connection.setPaintStyle({stroke: 'red', lineWidth: '10px'});
-              connection.addOverlay(["Label",{label:"Composition",location:0.5}]);
-              connection.addOverlay(["Diamond", {width: 15,length: 30,location: 1}]);
-              break;
-            //closed arrow, solid line
-            //generalization
-            case 'orange':
-              connection.removeAllOverlays();
-              connection.setPaintStyle({stroke: 'orange', lineWidth: '10px'});
-              connection.addOverlay(["Label",{label:"Generalization",location:0.5}]);
-              connection.addOverlay(["Arrow",{width: 15,lenght: 30,location:1}]);
-              break;
-            //closed arrow, dashed line
-            //realization
-            case 'yellow':
-              connection.removeAllOverlays();
-              connection.setPaintStyle({"dashstyle":'3 3',stroke:'yellow',strokeWidth:5});
-              connection.addOverlay(["Label",{label:"Realization",location:0.5}]);
-              connection.addOverlay(["Arrow",{width: 15,lenght: 30,location:1}]);
-              break;
-          }
-        });
-      }
-    });
+    
 
     
   }
@@ -358,7 +244,7 @@ export class ClassBoxComponent implements OnInit, AfterViewInit,DoCheck, AfterVi
             notAlreadyConnected = true;
           }
           if(notAlreadyConnected){
-            var connection = this.service.jsPlumbInstance.connect({uuids:[this.id+'_bottom',targetId]});
+            var connection = this.guiService.jsPlumbInstance.connect({uuids:[this.id+'_bottom',targetId]});
             //composition
             connection.removeAllOverlays();
             connection.setPaintStyle({stroke: 'red', lineWidth: '10px'});
@@ -416,7 +302,7 @@ export class ClassBoxComponent implements OnInit, AfterViewInit,DoCheck, AfterVi
   //edit name
   //GUI no tests
   editName(){
-    this.service.connectionsUpdateWrapper();
+    this.guiService.connectionsUpdateWrapper();
     //update connections with new name
     var connections = this.service.findClass(this.name)['connections'];
     for(var i = 0;i<connections.length;i++){
@@ -435,12 +321,12 @@ export class ClassBoxComponent implements OnInit, AfterViewInit,DoCheck, AfterVi
 
     var id = this.id;
     //reset id of endpoints on name change
-    this.service.jsPlumbInstance.selectEndpoints({source: element}).each(function(endpoint){
+    this.guiService.jsPlumbInstance.selectEndpoints({source: element}).each(function(endpoint){
       endpoint['elementId'] = id;
     });
 
     //update id jsPlumb stores
-    this.service.jsPlumbInstance.setIdChanged(oldId,this.id);
+    this.guiService.jsPlumbInstance.setIdChanged(oldId,this.id);
 
 
 
@@ -450,7 +336,7 @@ export class ClassBoxComponent implements OnInit, AfterViewInit,DoCheck, AfterVi
   //GUI no change
   deleteClass(){
     this.service.allClasses.splice(this.service.allClasses.indexOf(this.service.findClass(this.name)),1);
-    this.service.jsPlumbInstance.remove(this.id);
+    this.guiService.jsPlumbInstance.remove(this.id);
   }
 
   updateChip(){
@@ -460,16 +346,7 @@ export class ClassBoxComponent implements OnInit, AfterViewInit,DoCheck, AfterVi
     else{
       this.methods[this.methods.indexOf(this.oldChipValue)] = this.chipAttribute;
     }
-    // for(var i = 0;i<this.variables.length;i++){
-    //   if(this.variables[i] == this.oldChipValue){
-    //     this.variables[i] = this.chipAttribute;
-    //   }
-    // }
-    // for(var i = 0;i<this.methods.length;i++){
-    //   if(this.methods[i] == this.oldChipValue){
-    //     this.methods[i] = this.chipAttribute;
-    //   }
-    // }
+
     this.updateValues();
     this.edit = '';
     this.oldChipValue = ''
@@ -491,6 +368,7 @@ export class ClassBoxComponent implements OnInit, AfterViewInit,DoCheck, AfterVi
     this.methods = cls['methods'];
   }
 
+  //makes sure when changing name that the class with that name doesn't already exist
   existenceCheck(){
     if(this.editorName){
       var cls = document.querySelector('.'+this.editorName);
